Added example 2 component
