+++ src/app/screens/Dashboard/screens/UseLayoutEffect/components/Example2/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Example2() {
+  return (
+    <div className="column">
+      <h2>Example 2</h2>
+    </div>
+  );
+}
+
+export default Example2;
