+++ src/scss/components.scss
@@ -46,3 +46,7 @@
   max-height: 40px;
   padding: 8px;
 }
+
+.separator {
+  border-bottom: 1px solid $blue;
+}
